NC05-035 flow - Computed globals queried before computed
Rework Has_Output into Has_No_Output and clarify in the comments that
if we don't know for sure we say False (i.e. we might have some output).

We might not know for sure if we're in phase 1 and are building the flow
graphs for global generation. Assuming we might have some outputs is
the safe default anyway for Get_Abend_Kind.