NC12-017 flow - use enum instead of string for message tags
The various tags that a flow emitted message can have are now
enumerated in vc_kinds.ads. This will facilitate keeping the UG
updated whenever a new message kind is added. The categorization
of these can be improved in the future (we will have to think
of a way of splitting them into groups).