NC17-035 flow - check mode of prefixes of attribute Old
This patch adds a missing flow check. This check has to do with the
mode of variables that serve as prefixes of 'Old attributes. If such
a prefix is NOT an import then flow analysis issues a high check.

* flow-analysis / flow
(Get_Initial_Vertex): Function was moved from flow-analysis to flow
(Get_Final_Vertex): Function was moved from flow-analysis to flow

* flow-control_flow_graph
(Do_Postcondition): During the analysis step we also check that
  variables that serve as prefixes of attributes Old are imports.

* Testsuite
Added a test and updated one affected test