N514-049 flow - improve error message
When a hidden variable of a remote package is used by a subprogram
without having been mentioned in the subprogram's Global aspect we
issue an error saying that a Global aspect is required to make the
variable visible. The relevant message now also mentions the remote
hidden variable that caused the issue.