NC18-019 flow - analyze declaration of nested package bodies
Flow analysis used to always ignore the declarations section of nested
packages during the generation of flow graphs. This resulted in missing
variables in the case of nested packages with hidden state that was
not exposed though a state abstraction.

This has now been fixed. In the case of nested packages with hidden
inexposed state, flow analysis processes the declarations of the
nested package in a similar way to how it processes the visible
declarations of a nested package's spec.