O219-034 flow - report successful checks
Now that AA is a check we may continue on with analysis. This change
special cases out vertices for procedure calls in the DDG construction
to make them unaffected by their order.

This also fixed an obscure issue with a volatile with ER that is passed
to a procedure with an incorrect contract.

Test disabled for now since proof aborts - will raise a separate ticket
for that.