MA11-028 Add section on generation of refined data and flow dependencies
This section should be in "Precise Generation for |SPARK| Subprograms"
instead of "Auto Completion for Incomplete Contracts". Indeed, the latter
is about completion of contracts that does not require the implementation
of a subprogram, while the latter is about generation of contracts from
the implementation of a subprogram. Also add a test showing the expected
result (currently failing, see O303-022).