O305-024 Dynamic property of uninitialized
Do not assume the dynamic property of uninitialized
variables if the range is empty.

* gnat2why-expr.ads
(Assume_Dynamic_Property): Now takes an additionnal
parameter stating if the expression is supposed to be
initialized or not. If it is not then scalar
expressions are only assumed to be in range if the
range is not empty.