O318-021 Refactor scalar conversions and fix missing case
The function for type conversion between scalar types was getting too
complex to understand, and as a result a case was missing for converting
from a fixed-point type to a modular type. The steps in this function
have been reordered for clarity, pushing generation of range checks to
the function Do_Range_Check meant for this purpose. Clarification of
Do_Range_Check function is left for work on O318-035.

* gnat2why/why/why-gen-expr.adb
(Do_Range_Check): Add different cases for handling of modular types,
  previously in Insert_Scalar_Conversion.
(Insert_Scalar_Conversion): Clarify sequence of steps to convert from
  one scalar type to another.

* testsuite/gnatprove/tests/O318-021__modular/fpmod.adb
* testsuite/gnatprove/tests/O318-021__modular/test.opt
* testsuite/gnatprove/tests/O318-021__modular/test.out
Update test which passes now.