O205-003 missing dynamic property on refined post
When a subprogram has a refined post, the refined post
is the only thing which is known about the program when
proving the postcondition.
We should also assume the dynamic property of outputs
of the subprogram.

*  gnat2why/why/gnat2why-subprograms.adb
(Generate_VC_For_Subprogram): Assume dynamic property of
effects along with the refined postcondition.