O331-041 flow - spurious error about missing global
When an object who's type has a Default_Initial_Condition aspect
is declared, flow analysis checks for uninitialized variables
in the expression of the Default_Initial_Condition. At that point
any references to the Type are replaced with references to the
object. For records, we also need to replace components of the
type with components of the object (this was previously missing
and led to a spurious error about a missing global variable).