O306-021 Use globals mode for dynamic property of globals
When assuming the dynamic property of global objects
referenced in a procedure, we now assume initialization
for globals that are read (and therefore are not out).

* gnat2why/spark/spark_util.ads
(Is_Declared_In_Unit): Checks if a variable is declared
in a given subprogram/package.

* gnat2why/why/gnat2why-subprograms.adb
(Compute_Dynamic_Property_For_Inputs) Assume
initialization of read objects.