O401-043 Attribute Size always returns a non-negative value
The Why3 identifier attr__size renamed to value__size -- now we follow
the convetion of GNAT, which defines 'Value_Size and 'Object_Size (in
addition to 'Size defined by Ada).

For array and record types both value__size and object__size are defined
as uninterpreted functions. Axioms state that these functions always
return non-negative values. For other types we take the exact sizes
which are already known to the compiler.

* why-atree-modules.adb
  (Insert_Why_Symbols): attr__size renamed to value__size

* why-gen-names.ads
  (Why_Name_Enum): entries related to size attribute are now sorted
    alphabetically

* why-gen-names.adb
  (Insert_Why_Symbols): as above

* why-gen-records.adb
  (Declare_Attributes): generate axioms related to size attribute

* ada__model.mlw
  Axioms about the size attributes for arrays

* tests/O108-025__attribute_size/test.out
  Test output updated