O108-025 WNE_Attr_Size renamed to WNE_Attr_Object_Size
The meaning of attribute Size applied to objects and types is different.
This is now reflected in having separate WNE_Attr_Object_Size and
WNE_Attr_Value_Size insted of a single WNE_Attr_Size. ('Object_Size and
'Value_Size are already defined in GNAT so we follow the same convention).