O417-023 Range check on static subtype of a dynamic type
We used to not check the bounds of a scalar subtype when
they were static because the frontend was already doing
it. This is not true if the subtype's base type is
itself dynamic.

* gnat2why/why/gnat2why-expr.adb
(Case_Expr_Of_Ada_Node): Do not check the bounds of the
conditions as they are checked against the type's base
range if the type is dynamic.
(Check_Scalar_Range): Check the bounds of a static type
if the base type is dynamic.