O403-015 Translate dispatching equality
Equality on classwide types was translated as
equality on the specific types.
Dispatching equality is now translated as equality
on tags and a call to an uninterpreted function
__dispatch_eq.

* why-gen-records.adb
(Declare_Equality_Function): Declare an uniterpreted
dispatching equality on Root types.

* gnat2why-expr.adb
(New_Op_Expr): Translate equality on classwide using the
uninterpreted dispatching equality of the root type.

* ada__containers__formal_*
Do not assume why3 equality between elements from the
outside world and elements stored in the container.