O317-004 Proper treatment of volatile constants in proof
As part of the refactoring of utility units, we now detect that constants
of a volatile are not in SPARK, we issue an error when the query function
for flavors of volatile is called on a constant, and we generate a
variable of unknown type __private for variables not in SPARK (to be used
in effects).

* spark_definition.adb
(Mark_Object_Entity): Detect volatile constants as a violation of SPARK.

* spark_util.adb
(Has_Volatile_Flavor): Raise error if called on constant.

* gnat2why-decls.adb
(Translate_Variable): Generate simple variable declaration for variable
  not in SPARK, only useful for effects.

* gnat2why-util.adb
(Is_Mutable_In_Why): Only look at volatile aspect for objects that are not
  a constant in SPARK.

* why-gen-binders.adb
(Mk_Item_Of_Entity): Generate simple binder with type __private for
  variables not in SPARK.