O331-003 Use base type for def axioms of non recursive
For scalar types which are not modeled as their why rep
type in Why, using instead this rep type
facilitate the job of automatic provers. As it is potentially
incorrect if there can be a runtime error in the expression
function body, it should not be done if the function may be
recursive.

* gnat2why/why/gnat2why-subprograms.adb
(Translate_Expression_Function_Body): Use why rep type for
defining axiom of non-recursive expression functions.