N318-007 Rewrite section on loops in GNATprove-by-Example
This is a merge of the previous sections on "Loop Examples" and
"Loop Patterns" that reuses the notions of patterns previously defined,
and extends them beyond arrays to collections, using formal vectors and
lists. Most previous examples in these sections have not been reused, but
they are still in gnatprove_by_example/examples until we decide to move
them in the testsuite, or to delete them.

* gnatprove.rst
Remove use of section on "Loop Patterns". Keep file for now for easier
reference.

* tutorial.rst
Minor renaming.

* loop.rst
Rewrite section on loop examples, based on patterns.

* Makefile.examples
Separate proof for examples that require a longer timeout (60s).

* examples/*.adb
New loop examples.

* results/*.prove, results/*.flow
New oracles for the loop examples.