O513-003 Do not issue spurious warnings on justification pragmas
When analysis is requested for a specific subprogram or line, it is
expected that no check message is issued for other subprograms. In that
case, ignore the justifications in other subprograms. Justifications in
the same subprogram may still trigger unwanted warnings, to be dealt
with if this proves to be annoying.

* gnat2why-annotate.ads|b
(Mark_Pragma_Annotate): Ignore pragma Annotate if analysis is requested
  for another subprogram.

* testsuite/gnatprove/tests/O513-003__justification/test.opt
Test passes now.