O521-014 Remove restriction on calls to procedures marked No_Return
This lifts the SPARK RM restriction forcing calls to procedures marked
No_Return to appear either at the end of a subprogram or as the only
statement inside an if-statement without else/elsif part. Indeed, this
restriction is not needed anymore, similarly to what was done for
raise-statements already.

* docs/lrm/source/subprograms.rst
Remove SPARK RM legality rule.

* gnat2why/spark/spark_definition.adb
Remove checking of the SPARK RM legality rule.