O526-020 Force default timeout of 1s when none specified in command-line
GNATprove default value of timeout for provers is 1s, but this was not
properly enforced for provers other than Alt-Ergo (relying for Alt-Ergo on
its own default timeout of 1s, at least in our mirrored version). This is
now fixed.

* gnatprove/configuration.adb|s
(Read_Command_Line): Set initial value for Timeout to 1 (= 1 second).