O512-032 Remove most loop invariants by using for loops
Instead of using plain loops, use for loops so that GNATprove readily
knows the range of the iteration variable without the need for a loop
invariant to state this range. Also define constant values in terms of
the iteration variable, instead of doing manual optimization at code
level (the optimizer will do it at levels -O2 at least). As a result,
only 7 loop invariants remain, all of which deal with the frame condition,
i.e. parts of a record that are preserved by the loop (and thus could be
generated automatically one day).

As a result of these changes, an intermediate assertion is needed that is
not proved, on line 580, which restates the value given to a constant
on line 564 at declaration. To be investigated.