O512-032 Comment out contracts on internal subprograms
Contracts on internal subprograms are not needed in SPARK 2014, as
GNATprove can inline calls to subprograms without contracts. Leave them
only as comments so that people can choose to verify the three subprograms
Get_64_LSB_First, Put_64_LSB_First and Copy_Msg_To_B independently. But by
default these are verified inlined in their caller, and do not need the
user to write a contract for them.