O429-046 flow - generate globals for tasks
This patch extends support of globals generation to tasks.

* spark_util
(Is_Requested_Subprogram): Function renamed into
  Is_Requested_Subprogram_Or_Task and now works on tasks too

* spark_frame_conditions
Bypassing usage of computed globals for tasks (since this has
not been implemented yet). We consider the computed globals to
be completely null (no inputs and no outputs).

* spark_util
(Task_Body_Entity): New function that takes an E_Task_Type and
  returns the corresponding E_Task_Body (if it exists).

* flow_computed_globals
The gg information that we store in the ali files now also
says if the information relates to a subprogram or a task.

* Testsuite
(O429-046__flow_tasking_cfg): Added some more tests