O528-003 Mark functions with volatile param/return as not in SPARK
The rule that functions with volatile param/return are not in SPARK was
enforced in the frontend for code marked SPARK_Mode (On), but not in
gnat2why for code implicitly marked SPARK_Mode (Auto) for which violations
of the rule result in the marking of the function as not in SPARK. Now
fixed.

* gnat2why/spark/spark_definition.adb
(Mark_Function_Specification): Detect functions with effectively volatile
  parameter or return.
* testsuite/gnatprove/tests/O528-003__atomic_param
New test.