NB19-021 proof - upgrade to latest cvc4
Thanks to the new options (thanks goes to Liana!) we can control how
resource counting works and can now move to the latest CVC4 version.

Reworked the option string creation for CVC4 in gnatprove.

Also removed --int-when=full-last-call and --user-pat=trust, since they
are now the default.

Changed step limit fudge factor for CVC4 (used to be 10x, now
it is 5000 + 25x); this is because more activities inside CVC4
now count as steps.

Updated test output.