O523-003 Bounds forgotten after slice assignments
An assignment of a slice of an array is translated as
an abstract modification with a postcondition setting
the values of elements of the array appropriately.
If the array is not statically contrained nor in
split form, we must also assume that its bounds are
preserved.

* gnat2why-expr.adb
(One_Level_Update): Assume preservation of bounds of
arrays when assigning a slice.

* why-gen-arrays.ads
(New_Bounds_Equality): New function creating a
predicate to express preservation of array bounds.