O603-004 flow - volatile globals of nonvolatile functions
This patch extends the info that flow analysis stores on ALI files.
To be more precise, information about volatile variables and
external state abstractions is now also stored there. This info is
used to deduce if a magic string has volatile properties or not.

Additionally, a new procedure has been implemented that emits an error
whenever a nonvolatile procedure has a volatile global item (both a
variable or a state abstraction).

* vs_kinds
(Flow_Tag_Kind): Non_Volatile_Function_With_Volatile_Effects has been
  added to the type.

* flow_computed_globals
(GG_Is_Volatile): New function that checks if a magic string is a volatile

(GG_Has_Async_Readers): New function that checks if a magic string is a
  volatile with property Async_Readers set.

(GG_Has_Async_Writers): New function that checks if a magic string is a
  volatile with property Async_Writers set.

(GG_Has_Effective_Reads): New function that checks if a magic string is a
  volatile with property Effective_Reads set.

(GG_Has_Effective_Writes): New function that checks if a magic string is a
  volatile with property Effective_Writes set.

* flow_types
(Is_Volatile): Augmented so that it works on magic strings.

(Has_Async_Readers): Augmented so that it works on magic strings.

(Has_Async_Writers): Augmented so that it works on magic strings.

(Has_Effective_Reads): Augmented so that it works on magic strings.

(Has_Effective_Writes): Augmented so that it works on magic strings.

* flow-analysis
(Check_Function_For_Volatile_Effects): Checks that the subprogram does not
  have any volatile effects except if so specified. This check only does
  something when called on [generic] functions.

* Testsuite
(O603-004__flow_vol_globs_on_nonvol_func): New test that exercises the
  new functionality.