N816-002 Detect illegal uses of External_Axiomatization
The tool now does not simply assume SPARK_Mode Off on
private part and body of a package with external
axiomatization, it verifies it.

* spark_definition.adb
(Mark_Package_Body): Raise a violation if the body of
a package with external axioms is SPARK_Mode On.
(Mark_Package_Declaration): Raise a violation if the
private part of a package with external axioms is
SPARK_Mode On.

* why-atree-sprint.adb
(Print_Custom_Declaration): Use local name of an
unfound theory file in the error message as it is
more reproducible.