O110-002 Detect use of type name in default expressions
SPARK RM 3.8(2) states that the default expression of a component
declaration shall not contain a name denoting the current instance
of the enclosing type. This is now detected in GNATprove. Note that GNAT
currently rejects this at compilation, although it is valid Ada.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): Detect SPARK violation.

* O110-002__default_init
New test.