O612-018 rubbish switches not detected
gnatprove would not complain about unkown extra arguments. The reason
was that these were interpreted as files. Later, gnatprove would check
whether the files are spec, body or subunit/separate files. For some
reason, this test returns "subunit", and the gnatcoll API returns the
empty string when asking for the main unit of that subunit. This empty
string is then passed as "file name" to gprbuild, which does not
complain (doesn't even notice).

This is fixed by checking whether the "main unit" is an actual file,
just as we do when switching from spec to body.

configuration.adb:
(Sanitize_File_List) check whether returned file does exist