N318-011 Initial implementation of counter-examples.
If all provers fail to prove a VC, additional prover run is scheduled to
get a counter-example. This prover run is made using cvc4.
Note that cvc4 is run with different arguments than in case of standard
run and its input can also differ (there are some why3 transformations
done just in the case of generating counter-examples).
Counter-examples are currently supported only for variables of discrete
types.

To be able to generate counter-examples, special labels are inserted to
variables when translating spark code to why3 code and why3 location tags
are generated.

A counter-example is written to the file in gnatprove directory. This file
is reported in the *.spark file in gnatprove directory.

There is an option of gnatprove "--counter-example=(on|off)" to
enable/disable this functionality.

A query for a counter-example can cause cvc4 to output undesirable
warnings and errors (e.g., because cvc4 does not support non-linear
arithmetic when generating counter-ezamples). To make tests passing,
these errors and warnings are filtered out in
testsuite/gnatprove/lib/python/test_support.py.