O622-007 Use steps support of why3
The number of steps was passed to provers by writing the number to the
why.conf file generated by gnatprove. Additionally, step number needed
to be passed to why3 (why3 needs to know whether steps were given or not).
This commit removes the code for handling steps from gnatprove. Steps
are now just passed to gnatwhy3 and the support of steps built in why3
is used.

To use the support of steps built in why3, test.conf generated by gnatprove
must for each prover additionally contain a version of the command that
is used to run the prover in case that steps are passed to gnatwhy3
(command_steps).

* gnatrove/gnatprove.adb
(Generate_Why3_Conf_File): Removed code for adjusting the steplimit for
CVC4 (this is done in gnatwhy3 now). Removed writing of steplimit to
the command to run provers. Added command to run provers with steplimit.

* testsuite/gnatprove/tests/N804-036__bad_prover/test.conf
Adding command_steps to test.conf