O429-041 implement volatile functions for proof
This is achieved by generating a state which will be modified every time
the function is called.

Also, we need to remove the postcondition which states that the result
of the function call is equal to the logic function.

* gnatw2hy-subprograms.adb
(Generate_Subprogram_Program_Fun) add volatile state for a volatile
function, and remove the incorrect part of the postcondition