O311-023 flow - improve GG algorithm
This patch adds an extra graph that represents which local variables
can act as globals to which subprograms.

* flow-slice.adb
(Get_Local_Variables_And_Subprograms): Do not descend into nested
  subprograms.

* flow_generated_globals.adb
Added a new graph (Local_Graph) and then utilized it to prevent drawing
edges to local variables in the Global_Graph.

* graph.adb
(Close): When closing a graph the visitor procedure is now called after
  the edge has been added. This allows us to instantly remove the newly
  added edge if we want.