O609-013 Do not rename Why3 types
We plan to use type names to monomorphize reference types. For this
to work, we need types that must use compatible references to have
the same name.
Split forms of arrays and discrete types and statically constrained
array types are translated directly as maps and integer/bitvectors.

* gnat2why/why/why-gen-arrays.adb
(Declare_Constrained): Do not output a renaming for the map type.

* gnat2why/why/why-inter.adb
(New_Kind_Base_Type): Returns the why type to be used for E. It
will have the kind Kind but its name may be the same as another
type's name to avoid renaming. It is the case for split forms and
static array types.
(Eq_Base): Now checks the whole structure of the type id, not only
its name.
(Eq_In_Why): New equality function that only checks the names of
types.

* share/spark/theories/ada__model.mlw
(Discrete_Base)
(Unconstr_Array_*): Remove declaration of a __split type.