O403-005 nonprivate records can have a fullview not in spark
A record type can be derived from a tagged private extension
and therefore have its fullview not in SPARK.
Also fix some problems in spark_util's query functions for
records uncovered by the previous fix.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): Record types whose Etype have a fullview
not in SPARK have a fullview not in SPARK.

* gnat2why/spark/spark_util.adb
(Retysp): Do not go through Get_First_Ancestor_In_SPARK as it
stores Etypes of private types.
(Component_Is_Visible_In_SPARK): A component is visible in SPARK
if the fullview of the type in which it is defined is in SPARK.
(Search_Component_By_Name): If the record is a classwide record,
search in the Retysp of the type as the type itself may be
private.