O526-006 remove spurious warning about useless justification
There were sometimes warnings issued on useful pragma Annotate in
inlined subprograms, because either the original pragma did not
correspond to any message, or it happened that in *one* inlining the
pragma was indeed useless (but in others it might be useful).

We fix this by disabling such warnings on pragmas that are inlined or
from generic instances. This again is done in two steps:

* avoid marking pragmas in inlined subprograms
* avoid issuing warnings on pragmas that have chained locations

* gnat2why-annotate.adb
(Mark_Pragma_Annotate) skip pragma in inlined subprograms that are
  referenced
(Generate_Warning_On_Useless_Pragma) do not process pragmas with chained
  locations