O706-020 if expressions on floating point types
Alternatives of if expressions were converted to the
Why3 translation of the type of the expressions.
This seems to be unnecessary in general and harmful on
floating point types where it introduces a call to of_real.

* gnat2why/why/gnat2why-expr.adb
(Transform_Expr): Convert the alternatives of an if
expression to the expected type of the expression.