O707-018 Special case ASCII characters in Type_Of_Node
Entities for ASCII characters are not translated to Why,
and therefore not stored in the Symbol_Table.
The function returning the type of an entity should not
look in this table for ASCII characters but rather return
int directly.

* gnat2why/why/gnat2why-util.adb
(Why_Type_Of_Entity): Return EW_Int_Type on ASCII characters.