O407-016 Check tag on tagged conversion
If the type to which we are converting is not an
ancestor of the current type then we need to check
the object's tag.
For now, include a nearly always failing check as
a call to an uninterpreted predicate __compatible_tags.

* gnat2why/spark/spark_util.adb
(Check_Needed_On_Conversion): Also require check for
tagged conversion to a type which is not an ancestor.
(Is_Ancestor): Check if a type is an ancestor of
another type.

* gnat2why/why/why-atree-modules.ads
(M_Main): Store symbole for the __compatible_tags
predicate.

* gnat2why/why/why-gen-expr.adb
(Insert_Record_Conversion): Insert tag check when
necessary.

* gnat2why/why/why-gen-records.adb
(Insert_Tag_Check): Use __compatible_tags to insert
a tag check.

* share/spark/theories/_gnatprove_standard.mlw
(Main): Declare __compatible_tags. It is uninterpreted
but has a reflexivity axiom.