O601-022 Membership tests on specific tagged types
An expression is a member of a specific tagged type
if it has the correct tag and the correct discriminant.

* gnat2why/why/gnat2why-expr.adb
(Transform_Simple_Membership_Expression): For membership
in a specific tagged type, check that the object's tag
and the type's tag match.
(Compute_Default_Init): Assume the value of the tag of
default initialized tagged variables.