C17-015 Support membership test on classwide types
If the specific type associated to a classwide type
is constrained, we currently check the discriminant,
unlike what GNAT does. See O713-018.

* gnat2why/spark/spark_definition.adb
(Mark): Allow membership tests on classwide types.

* gnat2why/why/gnat2why-expr.adb
(Transform_Simple_Membership_Expression): A membership
test for inclusion into a classwide type checks that
the expression's tag is compatible with the type's tag.