O706-001 Z3 steps in config file.
The goal of this commit is to enable steps for Z3 in gnatprove. This means
to change the configuration in the why3.conf file which is generated by
gnatprove.adb and to remove a warning when z3 is called with steps.

* gnatprove/gnatprove.adb
(Generate_Z3_Setion): Add memory_max_alloc_count option to command_steps
command.
* gnatprove/configuration.adb
(Check_Prover_List): removed this procedure
(Prover_Is_Requested): removed this procedure
(Read_Command_Line): remove call to Check_Prover_List