O721-021 Duplicate generation of ref for cloned record
We should not generate a new ref type for a type which
is a clone of an existing one if the names coincide.

* gnat2why/why/why-gen-records.ads
(Record_Type_Is_Clone): New function to decide if a
record type is a clone or not.

* gnat2why/why/gnat2why-types.adb
(Translate_Type): Do not generate a new ref type for
cloned record types.