O713-018 Detect constrained classwide type in GNATprove
As it is unclear whether we should take into account the
values specified for their dicriminants.

* spark_definition.adb
(Mark_Type_Entity): Constrained subtypes of class-wide
types and class-wide types with a contrained specific
type are not supported yet.
Do not mark class-wide view of constrained record types
when marking their specific type.