N318-011 O619-018 Printing location tags for all constructs triggering VC.
When new location tag is printed, it is checked whether its location is
different from the location of the tag that was printed before. This
is not suitable for terms corresponding to constructs that trigger VC.
For these terms, location tag must be printed in any case. This is done
by this commit.

* gnat2why/why/why-atree-sprint.adb
(Print_Sloc_Tag): added parameter Force. When it is set to true, the
location tag is always printed.
(Print_Label): location tag is always printed if the node corresponds to
the construct that triggers VC.