O709-035 Adapt to new separation between constants/variables in xrefs
SPARK Xrefs now identify as constants only those constants with a static
initializing expression, to avoid miclassifying a constant with variable
input as constant. As a result, we need to adapt the translation of
external object to avoid treating objects in axiomatized units as
external.

* gnat2why/why/gnat2why-decls.adb
(Translate_External_Object): Ignore objects in axiomatized units.

* tests/O709-035__constant_with_var_input/test.out
Update test output.