O730-012 Logic functions for dynamic bounds of scalars
This is to improve readability of the Why3 generated
files. We declare functions for dynamic bounds of scalars
in their main module. They have all their variable inputs
as arguments. Their are defined using an axiom in the
type's axiom module.

* gnat2why/why/gnat2why-expr.adb
(New_Predicate_Call): New parameter params for transformation
parameters. It is used to determine if we should dereference
the arguments of the call.

* gnat2why/why/gnat2why-types.adb
(Create_Axioms_For_Scalar_Bounds): New function to emit
axioms for bounds in scalar types' axiom module.

* gnat2why/why/why-atree-modules.adb
(Insert_Why_Symbols): Insert symbols first and last on
dynamic scalar types.

* gnat2why/why/why-gen-binders.adb
(Get_Args_From_Expression): New function to get arguments
for a call to a function symbol representing an expression.
(Get_Args_From_Variables): New parameter Ref_Allowed to
decide if we should introduce dereferences.
(Get_Binders_From_Expression): New function to get the
binders that should be used for the arguments of a function
symbol representing an expression.
(Get_Binders_From_Variables): New parameter Compute to
specify if missing binders in the symbol table should be
computed. This is necessary when we use them to compute
the parameters of a function for an expression while in
the type modules.

* gnat2why/why/why-gen-expr.adb
(New_Attribute_Expr): Calls the predefined functions for
dynamic scalar bounds. New argument of Use_Pred to say
that the predefined function should not be used.

* gnat2why/why/why-gen-scalars.adb
(Define_Scalar_Attributes): Declare functions for dynamic
bounds of scalar types.