minor rename from Current to Direct
Rename from Current, which refers to time, to Direct, which refers to graph
adjacency (the opposite of Current is Future, which makes little sense;
the opposite of Direct is Indiderect, which is precisely what we compute
by closing the graph call).

* gnat2why/spark/spark_frame_conditions.ads
* gnat2why/spark/spark_frame_conditions.adb
* gnat2why/flow/flow.adb
(Collect_Current_Computed_Globals): now is Collect_Direct_Computed_Globals.