minor ALI parser refactor with GNAT.String_Split
The crude parser that search for names in a line of ALI file is now
implemented with a single call to GNAT.String_Split. This package is still
too heavy for our needs, but now at least we are not bothered with low-level
string processing.

* gnat2why/flow/flow_generated_globals.adb
(Get_Names_From_Line): use GNAT.String_Split instead of own parsing routine.