Minor refactoring
I noticed we have two ways about refering to the kind of thing analyzed:
in gg we have globals_origin_t, and in flow we use a subtype of
entity_kind. I have refactored this into a single new enum, which can
also be used as an array index (unlike the old one with the static
predicate). This stricter kind also removed a few "can't happen" when
other branches in case statements.

Also renamed the not-so-clear UG/FA/XR names to Origin_User, Origin_Flow,
and Origin_Frontend, which should make for more readable code; especially
since FA clashes with the use of "FA" throughout the rest of flow
analysis.