minor defence about ill-defined call
It is insane to ask whether Empty node is in analysed files or not; it
is better to forbid such questions with a precondition. This helps to
catch errors where we assumed that Empty node is indeed in analysed files.

* gnat2why/spark/spark_util.ads
(Is_In_Analyzed_Files): check if argument is not an empty node.