O811-006 few error msgs for PO not in SPARK
For protected body whose spec is not in SPARK give only one error message
instead of several for each of its subprogram.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): for task/protected body check if spec is in SPARK; it
not then give error and do not mark the body.
(Mark_Subprogram_Body): code moved to Mark_Type_Entity and is not executed
only once for type and not for each of its subprograms.

* testsuite/gnatprove/tests/O429-044__tasking_filter/
Test extended and output updated.