O429-048 fix comments for Is_Potentially_Blocking
All checks in Is_Potentially_Blocking refer to protected *type* and not
protected object. Instances of a protected type are not distinguished and
are all treated as a single protected object.

* gnat2why/flow/flow_generated_globals.adb
(Is_Potentially_Blocking): now comments reflect the fact that all checks
are done with a granularity of a protected type, not of a protected object.