O817-012 flow - rework ali format, parser and writer
Replace old ALI file IO with a new one using the serialisation package.
Not only is it shorter, it should also be less hassle to change in the
future. It is now also possible to read it without worrying about which
column we're in right now - this is not FORTRAN after all!