N318-011 Initial implementation of counterexamples for records.
To display content of records in counterexamples, records are projected
to their fields, which are either further projected (if they are of
type for which there is defined a projection) or displayed in the
counterexample.

This is achieved using the mechanism of projections supported by Why3.
For every record type, there are generated projections from values of this
type to values of types of the records fields (plus other projections,
because there are multiple why3 records generated for a single SPARK
record).
Then, in Why3, for every term that is labeled with the label
model_projected, there is created new term equal to the projection
of the term and the value of this term is displayed in the counterexample.
Note that projections are automatically composed and the term is always
projected with all maximal projections - i.e., when there is a projection
from P1 from A to B and projections P2 from B to C, and P3 from B to D,
the term of type A is projected using projection P3 = P1 P2 from A to C
and projection P4 = P1 P3 from A to D.

Definition of a projection consists of declaration of projection function,
which defines how value of the original type is projected to the value of
the resulting type and of declarations of metas that mark the function
as a projection function.

* common/vc_kinds.ads
Definition of the name of the meta for marking projection functions.

* gnat2why/why/gnat2why-decls.adb
(Translate_Variable): Adding counterexample labels for variables of record
type.

* gnat2why/why/gnat2why-types.adb
(Translate_Type): Emit declaration of projection from record type storing
a mutable field of type t to type t.

* gnat2why/why/gnat2why-util.adb
(Add_Counterexample_Labels): Add model_trace and model_projected labels
if the entity is of the record type.

* gnat2why/why/gnat2why.ads
Add documentation for meta model_projection, update documentation
for label model_trace.

* gnat2why/why/why-gen-decl.{ads|adb}
(Emit_Projection_Metas): Emit metas marking a function as a projection
  function.
(Emit_Record_Projection_Declaration): Emit declaration of projection
of values of given Why3 record type to given field of the type.

* gnat2why/why/why-gen-records.{ads|adb}
(Declare_Projection_Functions): Generate projection functions for SPARK
record type.
(Declare_Ada_Record): Added call to Declare_Projection_Functions.

* gnat2why/why/why-gen-scalars.adb
(Emit_Projection_Meta): Removed, the functionality moved to
Why-gen-decl.Emit_Projection_Metas.