N318-011 Counterexamples for records - displaying record fields.
Previously, record fields were displayed in counterexamples only for
record variables that were returned from functions. In other cases, record
fields were not displayed (only the names of enclosing variables were
displayed).

The reason was that why3 replaces record fields with simple variables with
same types as are the types of the fields. It propagated just names of the
enclosing variables (the variables of record types where the fields are
defined) to new variables, not names of the fields.

This commit solves this problem by adding labels of the form
"model_trace:name" to declarations of record fields. When why3 creates
new variable for record field, it uses this label to propagate the name
of the field to new variable.
Note that the name of the field can be "" meaning that the field should
not be displayed - just the name of the enclosing variable is displayed.
This is useful for fields that are not in SPARK source code, e.g,
__split_fields.

* gnat2why/why/why-sinfo.ads
* gnat2why/why/xgen/xtree_sinfo.adb
Adding labels to declaration of fields of record types.

* gnat2why/why/why-atree-sprint.adb
(Print_Record_Binder): Printing labels of of fields of record types.

* gnat2why/why/why-gen-binders.adb
(New_Constant_Record_Binders): Adding "model_trace:name" label the record
binder. "name" is the SPARK source name of the record field represented by
the binder.

* share/spark/theories/_gnatprove_standard.mlw
Adding label "model_trace:" to declarations of record fields that should
not be displayed in counterexample. In counterexample, accesses to these
fields will have the same names as accesses to enclosing variables
(variables of the record type where these fields are defined).

* why3
Traceability in counterexamples for record fields that are replaced by
eval_match by simple variables with the same name. Originally, all labels
from the variable that field was accessed were copied to new variables
representing fields of this variable. Therefore also
"model_trace:var_name" label was copied and thus the field had name
"var_name" in the counterexample.
Now names of the fields (or the names given in "model_trace:name" label
of the field declaration) are appended to "model_trace:*" label of new
variables representing record fields.